Extract debug button
