+++ site/src/pages/WorkspacePage/WorkspaceActions/DebugButton.stories.tsx
@@ -1,0 +1,54 @@
+import { Meta, StoryObj } from "@storybook/react";
+import { DebugButton } from "./DebugButton";
+import { MockWorkspace } from "testHelpers/entities";
+import { userEvent, waitFor, within } from "@storybook/test";
+
+const meta: Meta<typeof DebugButton> = {
+  title: "pages/WorkspacePage/DebugButton",
+  component: DebugButton,
+};
+
+export default meta;
+type Story = StoryObj<typeof DebugButton>;
+
+export const Default: Story = {};
+
+export const WithBuildParameters: Story = {
+  args: {
+    enableBuildParameters: true,
+    workspace: MockWorkspace,
+  },
+  parameters: {
+    queries: [
+      {
+        key: ["workspace", MockWorkspace.id, "parameters"],
+        data: { templateVersionRichParameters: [], buildParameters: [] },
+      },
+    ],
+  },
+};
+
+export const WithOpenBuildParameters: Story = {
+  args: {
+    enableBuildParameters: true,
+    workspace: MockWorkspace,
+  },
+  parameters: {
+    queries: [
+      {
+        key: ["workspace", MockWorkspace.id, "parameters"],
+        data: { templateVersionRichParameters: [], buildParameters: [] },
+      },
+    ],
+  },
+  play: async ({ canvasElement, step }) => {
+    const screen = within(canvasElement);
+
+    await step("open popover", async () => {
+      await userEvent.click(screen.getByTestId("build-parameters-button"));
+      await waitFor(() =>
+        expect(screen.getByText("Build Options")).toBeInTheDocument(),
+      );
+    });
+  },
+};
